Rename markdown wrapper script
